Move exceptions into its own subdirectory.
